pyproject.toml
[project]
name = "isle_chat"
version = "0.1.0"
description = "IsleChat — 基于 LangGraph 的长期陪伴型聊天 Agent，支持长短期记忆与可扩展技能系统"
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
requires-python = ">=3.12"
authors = [
    { name = "Jeff", email = "978814197@qq.com" },
]
keywords = [
    "agent",
    "chatbot",
    "langgraph",
    "langchain",
    "llm",
    "memory",
    "companion",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "langchain>=1.2.15",
    "langchain-openai>=1.1.12",
    "langgraph-checkpoint-postgres>=3.0.5",
    "psycopg[binary,pool]>=3.3.3",
    "pydantic-settings>=2.13.1",
]

[project.urls]
Homepage = "https://github.com/ynyg/IsleChat"
Repository = "https://github.com/ynyg/IsleChat"
Issues = "https://github.com/ynyg/IsleChat/issues"

[project.scripts]
isle-chat = "isle_chat:main"

[build-system]
requires = ["uv_build>=0.11.3,<0.12.0"]
build-backend = "uv_build"
